My version of the restAPI
This is my version of the restAPI which basically mirrors Gavin's as we worked on this together to some extent. We both pointed out helpful changes to each other so that this code will work for our situation. Since he uploaded first the only change I really contribute here happens to be a separate call in UserModel.php for a getRatings function. I also added .DS_Store to .gitignore since this is a Finder management file!